requirements.txt
fastapi>=0.95
uvicorn>=0.22
pydantic>=1.10
python-dotenv>=1.0
supabase>=2.0.0